move chain-related things into separate folder
