+++ slk/chain/xchain_transfer.py
@@ -2,7 +2,7 @@ import time
 
 from xrpl.models import Amount, Memo, Payment
 
-from slk.chain import Chain
+from slk.chain.chain import Chain
 from slk.common import Account
 from slk.sidechain_params import SidechainParams
