+++ slk/sidechain_repl.py
@@ -27,7 +27,7 @@ from xrpl.models import (
 )
 from xrpl.utils import drops_to_xrp
 
-from slk.chain import Chain, balances_data
+from slk.chain.chain import Chain, balances_data
 from slk.common import Account, same_amount_new_value
